final upload :D

+++ snowman.cpp
@@ -2,9 +2,10 @@
 #include <stdexcept>
 using namespace std;
 
+#include "snowman.hpp"
 
 namespace ariel {
-string snowman(long num){
+string snowman(int num){
     return "a";
 }
 }
