+++ snowman.hpp
@@ -1,4 +1,4 @@
 #include <string>
 using namespace std;
 
-namespace ariel {string snowman(long num);}
+namespace ariel {string snowman(int num);}
